Only purge comments that are pending. Grr....

+++ Gruta/Source/FS.pm
@@ -636,7 +636,7 @@ sub story_comments {
 			my $pf = $pend_path . join(':', ($topic_id, $story_id, $id));
 
 			# too old? delete
-			if (-M $f >= 7) {
+			if (-f $pf && -M $f >= 7) {
 				unlink $f;
 				unlink $pf;
 				next;
